Move authorization header generator helper to common place
This way it can be used from other test files too.
